Refactor and clean-up.
